+++ szconfigmgr/doc.go
@@ -1,4 +1,0 @@
-/*
-The szconfigmgr package is a wrapper over Senzing's G2Configmgr C binding.
-*/
-package szconfigmgr